requirements.txt
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.33.1
numpy==2.4.4
scipy==1.17.1
openai==1.54.0